+++ CHANGELOG.md
@@ -5,6 +5,8 @@
         - Clearer relocation options while you’re reporting a problem #2238
     - Admin improvements:
         - Allow moderation to potentially change category. #2320
+    - Open311 improvements:
+        - Fix bug in contact group handling. #2323
 
 * v2.4.2 (6th November 2018)
     - New features: